requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
pandas==2.0.3
schedule==1.2.0
xlsxwriter==3.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
Pyrogram==2.0.106
